color_eyre

Module config

Configuration options for customizing the behavior of the provided panic and error reporting hooks

Structs

EyreHook

An eyre reporting hook used to construct EyreHandlers

Frame

A representation of a Frame from a Backtrace or a SpanTrace

HookBuilder

Builder for customizing the behavior of the global panic and error report hooks

PanicHook

A panic reporting hook

PanicReport

A type representing an error report for a panic.

Theme

A struct that represents a theme that is used by color_eyre

Type Aliases

FilterCallback

Callback for filtering a vector of Frames