Struct color_eyre::config::EyreHook

pub struct EyreHook { /* fields omitted */ }

An eyre reporting hook used to construct EyreHandlers

Implementations

impl EyreHook

pub fn install(self) -> Result<(), InstallError>

Installs self as the global eyre handling hook via eyre::set_hook

pub fn into_eyre_hook(
self
) -> Box<dyn Fn(&(dyn Error + 'static)) -> Box<dyn EyreHandler> + Send + Sync + 'static>

Convert the self into the boxed type expected by eyre::set_hook.