color_eyre
Module section

Helpers for adding custom sections to error reports

Structs

IndentedSection

An indented section with a header for an error report

Traits

PanicMessage

Trait for printing a panic error message for the given PanicInfo

Section

A helper trait for attaching informational sections to error reports to be displayed after the chain of errors

SectionExt

Extension trait for constructing sections with commonly used formats